Store impersonated id in session rather than full user object
When using Doctrine ORM and a user entity with associated entities, serializing the user into the session causes some problems. The associations are not properly initialized when the user is unserialized in subsequent requests. This commit switches to just storing the user id rather than the whole object, which allows zfcuser's Storage/Db.php::read() to pull the user entity from the mapper, including properly set up association proxies.